requirements.txt
argcomplete==1.9.3
autopep8==1.3.3
awscli==1.16.14
awsebcli==3.14.8
base58==1.0.0
beautifulsoup4==4.6.1
boto3==1.9.4
botocore==1.12.4
bs4==0.0.1
cement==2.8.2
certifi==2017.7.27.1
cfn-flip==1.0.3
chardet==3.0.4
click==6.7
colorama==0.3.9
Django==2.1.2
djangorestframework==3.8.2
docutils==0.14
durationpy==0.5
future==0.16.0
hjson==3.0.1
idna==2.6
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
lxml==4.2.4
numpy==1.15.2
pandas==0.23.4
pathspec==0.5.5
placebo==0.8.2
pyasn1==0.4.4
pycodestyle==2.3.1
python-dateutil==2.6.1
python-http-client==3.1.0
python-slugify==1.2.4
pytz==2018.5
PyYAML==4.2b1
requests==2.21.0
rsa==3.4.2
s3transfer==0.1.13
selenium==3.7.0
semantic-version==2.5.0
sendgrid==5.6.0
six==1.11.0
sortedcontainers==2.0.5
termcolor==1.1.0
toml==0.9.6
tqdm==4.19.1
troposphere==2.3.3
Unidecode==1.0.22
urllib3==1.24.1
urlopen==1.0.0
virtualenv==16.2.0
Werkzeug==0.14.1
wsgi-request-logger==0.4.6
